Changes to the fips dataloader and examples

+++ odl/contrib/datasets/ct/examples/fips_reconstruct.py
@@ -4,7 +4,7 @@ import odl
 import odl.contrib.datasets.ct.fips as fips
 
 # Walnut example
-space = odl.uniform_discr([-20, -20], [20, 20], [2296, 2296])
+space = odl.uniform_discr([-20, -20], [20, 20], [2296, 2296], dtype='float32')
 geometry = odl.contrib.datasets.ct.fips.walnut_geometry()
 
 ray_transform = odl.applications.tomo.RayTransform(space, geometry)
@@ -14,7 +14,7 @@ data = fips.walnut_data()
 fbp_op(data).show('Walnut FBP reconstruction', clim=[0, 0.05])
 
 # Lotus root example
-space = odl.uniform_discr([-50, -50], [50, 50], [2240, 2240])
+space = odl.uniform_discr([-50, -50], [50, 50], [2240, 2240], dtype='float32')
 geometry = fips.lotus_root_geometry()
 
 ray_transform = odl.applications.tomo.RayTransform(space, geometry)
